pyproject.toml
# SPDX-FileCopyrightText: Copyright 2025 Arm Limited and/or its affiliates <open-source-office@arm.com>
# SPDX-License-Identifier: Apache-2.0
#
# Licensed under the Apache License v2.0
# See http://www.apache.org/licenses/LICENSE-2.0 for license information.

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61", "wheel"]

[project]
name = "pte-adapter-model-explorer"
version = "0.0.1"
description = "Adapter for Model Explorer to support PTE files for Ethos-U and VGF targets"
readme = "README.md"
requires-python = ">=3.10,<3.13"
keywords = ["model-explorer", "ethos", "pte", "visualization"]
license = { file = "LICENSE" }

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]

dependencies = [
  "ai-edge-model-explorer>=0.1.28,<0.2.0",
  "vgf-adapter-model-explorer>=0.4.0,<0.5.0",
  "tosa-adapter-model-explorer>=0.1.0,<0.2.0",
  "flatbuffers>25.9"
]

[project.urls]
Repository = "https://github.com/arm/pte-adapter-model-explorer"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-timeout",
    "ai-edge-model-explorer==0.1.28"
]
format = [
  "ruff==0.14.2"
]
coverage = [
  "pytest-cov"
]

[tool.pyright]
extraPaths = [
    ".",
    "./src",
]
autoSearchPaths = true
useLibraryCodeForTypes = true
diagnosticMode = "workspace"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"pte_adapter_model_explorer" = [
  "schema/*.fbs",
  "schema/LICENSE",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["src/pte_adapter_model_explorer/tests"]

[tool.ruff]
line-length = 79
target-version = "py312"
exclude = ["build", "src/pte_adapter_model_explorer/executorch_flatbuffer"]

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
select = [
    "F", #Pyflakes
    "B", #Bugbear
    "I", #Isort
    "SIM", #Simplify
    "C4", #Comprehensions
]

[tool.coverage.run]
omit = [
  "*/test_*",
  "*/config.py",
  "*/__init__.py",
  "*/executorch_flatbuffer/*",
]
